+++ be/src/main/java/com/example/be/controller/plan/PlanController.java
@@ -6,7 +6,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 import java.util.Map;
 
 @RestController
@@ -60,11 +59,5 @@ public class PlanController {
     @DeleteMapping("delete/{id}")
     public void delete(@PathVariable int id) {
         service.delete(id);
-    }
-
-    // 내 여행 달력으로 보기
-    @GetMapping("calendar")
-    public List<Plan> calendar() {
-        return service.calendar();
     }
 }
