+++ fe/src/page/plan/Plan.css
@@ -78,6 +78,16 @@ react-calendar__month-view__days__day--weekend {
 .react-calendar__tile--now {
     font-weight: bold;
     border-radius: 50px;
+    position: relative;
+}
+
+.react-calendar__tile--now::after {
+    position: absolute;
+    left: 0;
+    text-outline: 10px;
+    width: 10px;
+    height: 10px;
+    background: #51F8C4;
 }
 
 .react-calendar__tile--now:enabled:hover,
