+++ fe/src/page/plan/PlanList.jsx
@@ -11,8 +11,11 @@ import {
 import "react-calendar/dist/Calendar.css";
 import { FaRegQuestionCircle } from "react-icons/fa";
 import "./Plan.css";
+import Calendar from "react-calendar";
 
 function PlanList(props) {
+  const [filteredPlans, setFilteredPlans] = useState([]); // 필터링된 일정
+  const [selectedDate, setSelectedDate] = useState(null); // 선택된 날짜
   const [planList, setPlanList] = useState([]);
   const [addModalOpen, setAddModalOpen] = useState(false);
   const [search, setSearch] = useState({ type: "all", keyword: "" }); // 검색 정보 유지
@@ -31,6 +34,24 @@ function PlanList(props) {
         setCount(data.count);
       });
   }, [searchParams]);
+
+  // 선택된 날짜가 변경될 때 필터링
+  useEffect(() => {
+    if (selectedDate) {
+      const formattedDate = selectedDate.toISOString().split("T")[0]; // yyyy-MM-dd 형식
+      const filtered = planList.filter(
+        (plan) =>
+          formattedDate >= plan.startDate && formattedDate <= plan.endDate,
+      );
+      setFilteredPlans(filtered);
+    } else {
+      setFilteredPlans(planList); // 날짜 선택 안 하면 전체 일정 표시
+    }
+  }, [selectedDate, planList]);
+
+  const handleDateChange = (date) => {
+    setSelectedDate(date); // 선택된 날짜 업데이트
+  };
 
   useEffect(() => {
     const nextSearch = { ...search };
@@ -89,6 +110,28 @@ function PlanList(props) {
 
   return (
     <div className={"body"}>
+      <aside className={"calendar"}>
+        <Calendar
+          formatDay={(locale, date) =>
+            date.toLocaleString("en", { day: "numeric" })
+          }
+          showNeighboringMonth={false}
+          onChange={handleDateChange} // 날짜 선택 이벤트 핸들러
+          value={selectedDate} // 선택된 날짜 상태와 동기화
+          // 여행이 있는 날짜에 스타일 적용
+          tileContent={({ date }) => {
+            const formattedDate = date.toISOString().split("T")[0];
+            return planList.some(
+              (plan) =>
+                formattedDate >= plan.startDate &&
+                formattedDate <= plan.endDate,
+            ) ? (
+              <div className={"calendar-badge"}>여행</div>
+            ) : null;
+          }}
+        />
+      </aside>
+
       <button className="btn btn-dark" onClick={() => setAddModalOpen(true)}>
         새로운 여행 작성하기
       </button>
@@ -120,18 +163,18 @@ function PlanList(props) {
         </div>
       </div>
 
-      {planList.length === 0 ? (
+      {filteredPlans.length === 0 ? (
         <div className={"empty-container"}>
           <p>
             <FaRegQuestionCircle className={"empty-container-icon"} />
           </p>
           <p className={"empty-container-title"}>검색 결과가 없습니다.</p>
           <p className={"empty-container-description"}>
-            다른 검색어로 검색해주세요.
+            다른 날짜를 선택해주세요.
           </p>
         </div>
       ) : (
-        planList.map((plan) => (
+        filteredPlans.map((plan) => (
           <div key={plan.id}>
             <Card.Root flexDirection="row" overflow="hidden" maxW="xl" my={5}>
               <Box>
@@ -140,9 +183,7 @@ function PlanList(props) {
                   <Card.Description>{plan.description}</Card.Description>
                   <HStack mt="5">
                     <Badge>{plan.destination}</Badge>
-                    <Badge>
-                      {plan.startDate} ~ {plan.endDate}
-                    </Badge>
+                    <Badge>{plan.due}</Badge>
                   </HStack>
                 </Card.Body>
